+++ packages/react/src/internals/constants.tsx
@@ -1,0 +1,3 @@
+export const CACHE_KEY_INFINITE_QUERY = 'TRPC_INFINITE_QUERY' as const;
+export const CACHE_KEY_LIVE_QUERY = 'TRPC_LIVE_QUERY' as const;
+export const CACHE_KEY_QUERY = 'TRPC_QUERY' as const;
