+++ packages/react/src/internals/getCacheKey.tsx
@@ -1,0 +1,10 @@
+export function getCacheKey<TTuple extends [string, ...unknown[]]>(
+  [path, input]: TTuple,
+  extras?: string,
+) {
+  const cacheKey = [path, input ?? null];
+  if (extras) {
+    cacheKey.push(extras);
+  }
+  return cacheKey;
+}
